+++ src/app.ts
@@ -6,7 +6,6 @@ import timeout from "connect-timeout";
 import passport from "passport";
 import CONFIG from "./config";
 import { AuthServices } from "./components/auth/services";
-import { expressPinoConsoleLogger } from "./helpers/loggers";
 import * as errorHandler from "@/middlewares/errorHandler";
 import mainRouter from "@/routes";
 
@@ -28,7 +27,6 @@ export const createApp = (): express.Application => {
 
   if (CONFIG.APP.ENV !== "test") {
     app.use(morgan("dev"));
-    app.use(expressPinoConsoleLogger());
   }
 
   app.use(timeout(CONFIG.SERVER.TIMEOUT));
@@ -38,12 +36,10 @@ export const createApp = (): express.Application => {
   app.use(`/app/${CONFIG.APP.VER}`, mainRouter);
 
   // Error Middleware
-  for (const handler in errorHandler) {
-    if (handler === "notFoundError" || handler === "genericErrorHandler")
-      continue;
-    // @ts-ignore
-    app.use(errorHandler[handler]);
-  }
+  app.use(errorHandler.requestTimeoutHandler);
+  app.use(errorHandler.customErrorHandler);
+  app.use(errorHandler.authenticationErrorHandler);
+  app.use(errorHandler.prismaErrorHandler);
   app.use(errorHandler.genericErrorHandler);
   app.use(errorHandler.notFoundError);