+++ src/components/transactions/validators.ts
@@ -7,6 +7,10 @@ export const validateCashoutRequest = () =>
       isFloat: true,
       isEmpty: false,
       errorMessage: "amount is required",
+      custom: {
+        options: (value) => value > 0 && value < 2 ** 32,
+        errorMessage: "amount must be a number between 0 and 2**32",
+      },
     },
     bank_account: {
       in: ["body"],