+++ src/index.ts
@@ -3,11 +3,11 @@
 require("module-alias").addAlias("@", __dirname);
 
 import { createApp } from "./app";
-import { startServer, startWhatsappBot } from "./server";
+import { startServer } from "./server";
 
-if (process.env.NODE_ENV === "production") {
-  startWhatsappBot();
-}
+// if (process.env.NODE_ENV === "production") {
+//   startWhatsappBot();
+// }
 
 if (process.env.NODE_ENV !== "test") {
   const app = createApp();